Moved existing paginator adapters that are related with visits to the Visits namespace
