+++ module/Core/src/Visit/Paginator/Adapter/TagVisitsPaginatorAdapter.php
@@ -2,9 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Shlinkio\Shlink\Core\Paginator\Adapter;
+namespace Shlinkio\Shlink\Core\Visit\Paginator\Adapter;
 
 use Shlinkio\Shlink\Core\Model\VisitsParams;
+use Shlinkio\Shlink\Core\Paginator\Adapter\AbstractCacheableCountPaginatorAdapter;
 use Shlinkio\Shlink\Core\Repository\VisitRepositoryInterface;
 use Shlinkio\Shlink\Core\Visit\Persistence\VisitsCountFiltering;
 use Shlinkio\Shlink\Core\Visit\Persistence\VisitsListFiltering;
