+++ module/Core/test/Visit/Paginator/Adapter/VisitsForTagPaginatorAdapterTest.php
@@ -2,21 +2,20 @@
 
 declare(strict_types=1);
 
-namespace ShlinkioTest\Shlink\Core\Paginator\Adapter;
+namespace ShlinkioTest\Shlink\Core\Visit\Paginator\Adapter;
 
 use PHPUnit\Framework\TestCase;
 use Prophecy\PhpUnit\ProphecyTrait;
 use Prophecy\Prophecy\ObjectProphecy;
 use Shlinkio\Shlink\Common\Util\DateRange;
-use Shlinkio\Shlink\Core\Model\ShortUrlIdentifier;
 use Shlinkio\Shlink\Core\Model\VisitsParams;
-use Shlinkio\Shlink\Core\Paginator\Adapter\ShortUrlVisitsPaginatorAdapter;
 use Shlinkio\Shlink\Core\Repository\VisitRepositoryInterface;
+use Shlinkio\Shlink\Core\Visit\Paginator\Adapter\TagVisitsPaginatorAdapter;
 use Shlinkio\Shlink\Core\Visit\Persistence\VisitsCountFiltering;
 use Shlinkio\Shlink\Core\Visit\Persistence\VisitsListFiltering;
 use Shlinkio\Shlink\Rest\Entity\ApiKey;
 
-class VisitsPaginatorAdapterTest extends TestCase
+class VisitsForTagPaginatorAdapterTest extends TestCase
 {
     use ProphecyTrait;
 
@@ -34,8 +33,8 @@ class VisitsPaginatorAdapterTest extends TestCase
         $limit = 1;
         $offset = 5;
         $adapter = $this->createAdapter(null);
-        $findVisits = $this->repo->findVisitsByShortCode(
-            ShortUrlIdentifier::fromShortCodeAndDomain(''),
+        $findVisits = $this->repo->findVisitsByTag(
+            'foo',
             new VisitsListFiltering(DateRange::emptyInstance(), false, null, $limit, $offset),
         )->willReturn([]);
 
@@ -52,8 +51,8 @@ class VisitsPaginatorAdapterTest extends TestCase
         $count = 3;
         $apiKey = ApiKey::create();
         $adapter = $this->createAdapter($apiKey);
-        $countVisits = $this->repo->countVisitsByShortCode(
-            ShortUrlIdentifier::fromShortCodeAndDomain(''),
+        $countVisits = $this->repo->countVisitsByTag(
+            'foo',
             new VisitsCountFiltering(DateRange::emptyInstance(), false, $apiKey->spec()),
         )->willReturn(3);
 
@@ -64,13 +63,13 @@ class VisitsPaginatorAdapterTest extends TestCase
         $countVisits->shouldHaveBeenCalledOnce();
     }
 
-    private function createAdapter(?ApiKey $apiKey): ShortUrlVisitsPaginatorAdapter
+    private function createAdapter(?ApiKey $apiKey): TagVisitsPaginatorAdapter
     {
-        return new ShortUrlVisitsPaginatorAdapter(
+        return new TagVisitsPaginatorAdapter(
             $this->repo->reveal(),
-            new ShortUrlIdentifier(''),
+            'foo',
             VisitsParams::fromRawData([]),
-            $apiKey?->spec(),
+            $apiKey,
         );
     }
 }
