+++ module/Core/test/Visit/Paginator/Adapter/VisitsPaginatorAdapterTest.php
@@ -2,20 +2,21 @@
 
 declare(strict_types=1);
 
-namespace ShlinkioTest\Shlink\Core\Paginator\Adapter;
+namespace ShlinkioTest\Shlink\Core\Visit\Paginator\Adapter;
 
 use PHPUnit\Framework\TestCase;
 use Prophecy\PhpUnit\ProphecyTrait;
 use Prophecy\Prophecy\ObjectProphecy;
 use Shlinkio\Shlink\Common\Util\DateRange;
+use Shlinkio\Shlink\Core\Model\ShortUrlIdentifier;
 use Shlinkio\Shlink\Core\Model\VisitsParams;
-use Shlinkio\Shlink\Core\Paginator\Adapter\TagVisitsPaginatorAdapter;
 use Shlinkio\Shlink\Core\Repository\VisitRepositoryInterface;
+use Shlinkio\Shlink\Core\Visit\Paginator\Adapter\ShortUrlVisitsPaginatorAdapter;
 use Shlinkio\Shlink\Core\Visit\Persistence\VisitsCountFiltering;
 use Shlinkio\Shlink\Core\Visit\Persistence\VisitsListFiltering;
 use Shlinkio\Shlink\Rest\Entity\ApiKey;
 
-class VisitsForTagPaginatorAdapterTest extends TestCase
+class VisitsPaginatorAdapterTest extends TestCase
 {
     use ProphecyTrait;
 
@@ -33,8 +34,8 @@ class VisitsForTagPaginatorAdapterTest extends TestCase
         $limit = 1;
         $offset = 5;
         $adapter = $this->createAdapter(null);
-        $findVisits = $this->repo->findVisitsByTag(
-            'foo',
+        $findVisits = $this->repo->findVisitsByShortCode(
+            ShortUrlIdentifier::fromShortCodeAndDomain(''),
             new VisitsListFiltering(DateRange::emptyInstance(), false, null, $limit, $offset),
         )->willReturn([]);
 
@@ -51,8 +52,8 @@ class VisitsForTagPaginatorAdapterTest extends TestCase
         $count = 3;
         $apiKey = ApiKey::create();
         $adapter = $this->createAdapter($apiKey);
-        $countVisits = $this->repo->countVisitsByTag(
-            'foo',
+        $countVisits = $this->repo->countVisitsByShortCode(
+            ShortUrlIdentifier::fromShortCodeAndDomain(''),
             new VisitsCountFiltering(DateRange::emptyInstance(), false, $apiKey->spec()),
         )->willReturn(3);
 
@@ -63,13 +64,13 @@ class VisitsForTagPaginatorAdapterTest extends TestCase
         $countVisits->shouldHaveBeenCalledOnce();
     }
 
-    private function createAdapter(?ApiKey $apiKey): TagVisitsPaginatorAdapter
+    private function createAdapter(?ApiKey $apiKey): ShortUrlVisitsPaginatorAdapter
     {
-        return new TagVisitsPaginatorAdapter(
+        return new ShortUrlVisitsPaginatorAdapter(
             $this->repo->reveal(),
-            'foo',
+            new ShortUrlIdentifier(''),
             VisitsParams::fromRawData([]),
-            $apiKey,
+            $apiKey?->spec(),
         );
     }
 }
